Move azure boards service and test to new folder
