+++ src/issue-filing/services/azure-boards/azure-boards-issue-filing-service.tsx
@@ -2,8 +2,9 @@
 // Licensed under the MIT License.
 import { isEmpty } from 'lodash';
 import { IDropdownOption } from 'office-ui-fabric-react/lib/Dropdown';
-import { createSettingsGetter } from '../common/create-settings-getter';
-import { IssueFilingService } from '../types/issue-filing-service';
+
+import { createSettingsGetter } from '../../common/create-settings-getter';
+import { IssueFilingService } from '../../types/issue-filing-service';
 import { AzureBoardsSettingsForm } from './azure-boards-settings-form';
 import { azureBoardsIssueFilingUrlProvider } from './create-azure-boards-issue-filing-url';
