+++ src/issue-filing/services/azure-boards/create-azure-boards-issue-filing-url.ts
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { EnvironmentInfo } from '../../common/environment-info-provider';
-import { CreateIssueDetailsTextData } from '../../common/types/create-issue-details-text-data';
-import { title } from '../../content/strings/application';
-import { createIssueDetailsBuilder } from '../common/create-issue-details-builder';
-import { HTTPQueryBuilder } from '../common/http-query-builder';
-import { IssueDetailsBuilder } from '../common/issue-details-builder';
-import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from '../common/issue-filing-url-string-utils';
-import { HTMLFormatter } from '../common/markup/html-formatter';
+import { EnvironmentInfo } from '../../../common/environment-info-provider';
+import { CreateIssueDetailsTextData } from '../../../common/types/create-issue-details-text-data';
+import { title } from '../../../content/strings/application';
+import { createIssueDetailsBuilder } from '../../common/create-issue-details-builder';
+import { HTTPQueryBuilder } from '../../common/http-query-builder';
+import { IssueDetailsBuilder } from '../../common/issue-details-builder';
+import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from '../../common/issue-filing-url-string-utils';
+import { HTMLFormatter } from '../../common/markup/html-formatter';
 import { AzureBoardsIssueFilingSettings } from './azure-boards-issue-filing-service';
 
 const buildTags = (createIssueData: CreateIssueDetailsTextData, standardTags: string[]): string => {
