+++ src/tests/unit/tests/issue-filing/services/azure-boards/azure-boards-settings-form.test.tsx
@@ -6,15 +6,15 @@ import { TextField } from 'office-ui-fabric-react/lib/TextField';
 import * as React from 'react';
 import { IMock, Mock, Times } from 'typemoq';
 
-import { SettingsDeps } from '../../../../../DetailsView/components/settings-panel/settings/settings-props';
+import { SettingsDeps } from '../../../../../../DetailsView/components/settings-panel/settings/settings-props';
+import { OnPropertyUpdateCallback } from '../../../../../../issue-filing/components/issue-filing-settings-container';
 import {
     AzureBoardsIssueFilingService,
     AzureBoardsIssueFilingSettings,
-} from '../../../../../issue-filing/azure-boards/azure-boards-issue-filing-service';
-import { AzureBoardsSettingsForm } from '../../../../../issue-filing/azure-boards/azure-boards-settings-form';
-import { OnPropertyUpdateCallback } from '../../../../../issue-filing/components/issue-filing-settings-container';
-import { SettingsFormProps } from '../../../../../issue-filing/types/settings-form-props';
-import { EventStubFactory } from '../../../common/event-stub-factory';
+} from '../../../../../../issue-filing/services/azure-boards/azure-boards-issue-filing-service';
+import { AzureBoardsSettingsForm } from '../../../../../../issue-filing/services/azure-boards/azure-boards-settings-form';
+import { SettingsFormProps } from '../../../../../../issue-filing/types/settings-form-props';
+import { EventStubFactory } from '../../../../common/event-stub-factory';
 
 describe('AzureBoardsSettingsForm', () => {
     let props: SettingsFormProps<AzureBoardsIssueFilingSettings>;
