+++ src/tests/unit/tests/issue-filing/services/azure-boards/create-azure-boards-issue-filing-url.test.ts
@@ -1,15 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
 import { IMock, Mock } from 'typemoq';
-import { EnvironmentInfo } from '../../../../../common/environment-info-provider';
-import { title } from '../../../../../content/strings/application';
-import { AzureBoardsIssueFilingSettings } from '../../../../../issue-filing/azure-boards/azure-boards-issue-filing-service';
-import { createAzureBoardsIssueFilingUrlProvider } from '../../../../../issue-filing/azure-boards/create-azure-boards-issue-filing-url';
-import { HTTPQueryBuilder } from '../../../../../issue-filing/common/http-query-builder';
-import { IssueDetailsBuilder } from '../../../../../issue-filing/common/issue-details-builder';
-import { IssueUrlCreationUtils } from '../../../../../issue-filing/common/issue-filing-url-string-utils';
-import { IssueFilingUrlProvider } from '../../../../../issue-filing/types/issue-filing-service';
+
+import { EnvironmentInfo } from '../../../../../../common/environment-info-provider';
+import { title } from '../../../../../../content/strings/application';
+import { HTTPQueryBuilder } from '../../../../../../issue-filing/common/http-query-builder';
+import { IssueDetailsBuilder } from '../../../../../../issue-filing/common/issue-details-builder';
+import { IssueUrlCreationUtils } from '../../../../../../issue-filing/common/issue-filing-url-string-utils';
+import { AzureBoardsIssueFilingSettings } from '../../../../../../issue-filing/services/azure-boards/azure-boards-issue-filing-service';
+import { createAzureBoardsIssueFilingUrlProvider } from '../../../../../../issue-filing/services/azure-boards/create-azure-boards-issue-filing-url';
+import { IssueFilingUrlProvider } from '../../../../../../issue-filing/types/issue-filing-service';
 
 describe('createAzureBoardsIssueFilingUrl', () => {
     const testIssueDetails = 'html issue details';
